% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-doc.R
\docType{package}
\name{DesiredGainR-package}
\alias{DesiredGainR}
\alias{DesiredGainR-package}
\title{DesiredGainR: Auditable Multi-trait Selection Indices}
\description{
Package providing tools for:
\itemize{
\item desired-gain selection index optimization,
\item quadratic genomic selection index computation from trait genomic
estimated breeding values, explicit economic weights, genomic covariance
estimation, and published response diagnostics,
\item comparison of DG and QGSI outputs,
\item a high-level wrapper for full pipeline execution.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FAkohoue/DesiredGainR}
  \item \url{https://FAkohoue.github.io/DesiredGainR/}
  \item Report bugs at \url{https://github.com/FAkohoue/DesiredGainR/issues}
}

}
\author{
\strong{Maintainer}: Félicien Akohoue \email{akohoue.f@gmail.com} (\href{https://orcid.org/0000-0002-2160-0182}{ORCID})

Authors:
\itemize{
  \item Félicien Akohoue \email{akohoue.f@gmail.com} (\href{https://orcid.org/0000-0002-2160-0182}{ORCID})
}

}
\keyword{internal}
