% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_index.R
\name{index_information_efficiency}
\alias{index_information_efficiency}
\title{Cunningham information-deletion efficiency}
\usage{
index_information_efficiency(index)
}
\arguments{
\item{index}{A Smith-Hazel index or a general economic index.}
}
\value{
A data.table with one row per information variable.
}
\description{
A value of 0.98 means that the index after deleting one information variable
retains 98 percent of the full optimum-index standard deviation.
}
\references{
Cunningham EP (1969). The relative efficiencies of selection indexes.
\emph{Acta Agriculturae Scandinavica} 19:45-48.
}
