% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_qgsi.R
\name{print.quadratic_genomic_index}
\alias{print.quadratic_genomic_index}
\title{Print a quadratic genomic selection index}
\usage{
\method{print}{quadratic_genomic_index}(x, ...)
}
\arguments{
\item{x}{A \code{quadratic_genomic_index} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a quadratic genomic selection index
}
\keyword{internal}
